suggestions/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.2
watchdog==6.0.0
openai==0.27.0

